fix: install npiperelay by setup script.

+++ dot_local/bin/setup-socat-ssh-agent.sh
@@ -2,27 +2,19 @@
 
 NPIPERELAY=npiperelay.exe
 
-install_npiperelay() {
-    local NPIPERELAY_ZIP=https://github.com/jstarks/npiperelay/releases/download/v0.1.0/npiperelay_windows_amd64.zip
-
-    mkdir -p ~/.local/bin && cd ~/.local/bin
-    # unzip doesnt support to read from stdin.
-    curl -sL -o npiperelay.zip $NPIPERELAY_ZIP
-    unzip -o npiperelay.zip $NPIPERELAY
-    rm npiperelay.zip
-    cd -
-}
-
 start_ssh_agent() {
     export SSH_AUTH_SOCK=$HOME/.ssh/agent.sock
 
-    command -v $NPIPERELAY &> /dev/null || install_npiperelay
+    command -v $NPIPERELAY &> /dev/null || echo "WARN: not found ${NPIPERELAY}" >&2 && return
 
     if ! ss -a | grep -q $SSH_AUTH_SOCK; then
         rm -f $SSH_AUTH_SOCK
-        # prerequisite
-        # - Start-Service ssh-agent on Windows
-        # - ssh-add key on Windows
+        cat << 'EOF'
+Prerequisite for ssh-agent with npiperelay below
+- Start-Service ssh-agent on Windows
+- ssh-add key on Windows
+EOF
+
         (setsid socat UNIX-LISTEN:$SSH_AUTH_SOCK,fork EXEC:"${NPIPERELAY} -ei -s -v //./pipe/openssh-ssh-agent",nofork &) &> /dev/null
     fi
 }
